split batch container
